requirements.txt
# LangChain Core
langchain-community
langchain
langchain-core

# OpenAI
langchain-openai
openai

# Anthropic
langchain-anthropic

# Google Genimi (PaLM)
langchain-google-genai
google-generativeai

# Hugging Face
langchain-huggingface
transformers
huggingface-hub

# ollama
langchain-ollama

# .Evn support
python-dotenv

# ML Utilities
numpy
scikit-learn
ipykernel

# pytorch & TensorFlow
torch
tensorflow-cpu